refactor stream tests to set up shared servers
