+++ staging/src/k8s.io/apiserver/pkg/util/proxy/streamtranslator_test.go
@@ -23,13 +23,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"log"
 	"math"
 	mrand "math/rand"
 	"net/http"
-	"net/http/httptest"
 	"net/url"
-	"os"
 	"reflect"
 	"strings"
 	"sync"
@@ -51,56 +48,6 @@ import (
 	"k8s.io/component-base/metrics/legacyregistry"
 	"k8s.io/component-base/metrics/testutil"
 )
-
-var (
-	// spdyServer is the upstream SPDY server that the StreamTranslatorHandler connects to.
-	// It is shared across all tests in this package.
-	spdyServer *httptest.Server
-	// spdyServerMux is the router for the spdyServer. Tests register their specific handlers here.
-	spdyServerMux *http.ServeMux
-	spdyServerURL *url.URL
-
-	// streamTranslatorServer is the server that exposes the StreamTranslatorHandler.
-	// Test clients connect to this server. It is shared across all tests.
-	streamTranslatorServer *httptest.Server
-	// streamTranslatorServerMux is the router for the streamTranslatorServer.
-	// Tests register their specific StreamTranslatorHandler configurations here.
-	streamTranslatorServerMux *http.ServeMux
-	streamTranslatorServerURL *url.URL
-)
-
-// TestMain sets up the shared SPDY and StreamTranslator servers for the entire test suite.
-// This avoids the overhead of creating new servers for each test and eliminates race conditions
-// related to server startup and shutdown.
-func TestMain(m *testing.M) {
-	metrics.Register()
-
-	// Upstream SPDY server setup
-	spdyServerMux = http.NewServeMux()
-	spdyServer = httptest.NewServer(spdyServerMux)
-	var err error
-	spdyServerURL, err = url.Parse(spdyServer.URL)
-	if err != nil {
-		log.Fatalf("Failed to parse SPDY server URL: %v", err)
-	}
-
-	// StreamTranslator server setup
-	streamTranslatorServerMux = http.NewServeMux()
-	streamTranslatorServer = httptest.NewServer(streamTranslatorServerMux)
-	streamTranslatorServerURL, err = url.Parse(streamTranslatorServer.URL)
-	if err != nil {
-		log.Fatalf("Failed to parse StreamTranslator server URL: %v", err)
-	}
-
-	// Run tests
-	exitCode := m.Run()
-
-	// Teardown
-	spdyServer.Close()
-	streamTranslatorServer.Close()
-
-	os.Exit(exitCode)
-}
 
 // TestStreamTranslator_LoopbackStdinToStdout returns random data sent on the client's
 // STDIN channel back onto the client's STDOUT channel.
